Update ReformatR.R
In response to an issue regarding how the program and R handle CSV input files inconsistently (ignoring empty rows as if they didn't exist for one file, while retaining them in others) or according to a pattern not yet understood, I have made an update that mandates the indication whether empty rows are ignored or not. Additions to the code allow ReformatR to do its reformatting correctly based on this information.